refactor: extract center stats accumulator
