+++ src/stats/calculators/center.rs
@@ -28,83 +28,6 @@ pub struct CenterEvent {
     pub lateral_centering_distance: f32,
 }
 
-#[derive(Debug, Clone, Default, PartialEq, Serialize, Deserialize, ts_rs::TS)]
-#[ts(export)]
-pub struct CenterPlayerStats {
-    pub count: u32,
-    pub total_ball_travel_distance: f32,
-    pub total_ball_advance_distance: f32,
-    pub total_lateral_centering_distance: f32,
-    pub longest_center_distance: f32,
-    pub is_last_center: bool,
-    pub last_center_time: Option<f32>,
-    pub last_center_frame: Option<usize>,
-    pub time_since_last_center: Option<f32>,
-    pub frames_since_last_center: Option<usize>,
-}
-
-impl CenterPlayerStats {
-    pub fn average_ball_travel_distance(&self) -> f32 {
-        if self.count == 0 {
-            0.0
-        } else {
-            self.total_ball_travel_distance / self.count as f32
-        }
-    }
-
-    pub fn average_ball_advance_distance(&self) -> f32 {
-        if self.count == 0 {
-            0.0
-        } else {
-            self.total_ball_advance_distance / self.count as f32
-        }
-    }
-
-    pub fn average_lateral_centering_distance(&self) -> f32 {
-        if self.count == 0 {
-            0.0
-        } else {
-            self.total_lateral_centering_distance / self.count as f32
-        }
-    }
-}
-
-#[derive(Debug, Clone, Default, PartialEq, Serialize, Deserialize, ts_rs::TS)]
-#[ts(export)]
-pub struct CenterTeamStats {
-    pub count: u32,
-    pub total_ball_travel_distance: f32,
-    pub total_ball_advance_distance: f32,
-    pub total_lateral_centering_distance: f32,
-    pub longest_center_distance: f32,
-}
-
-impl CenterTeamStats {
-    pub fn average_ball_travel_distance(&self) -> f32 {
-        if self.count == 0 {
-            0.0
-        } else {
-            self.total_ball_travel_distance / self.count as f32
-        }
-    }
-
-    pub fn average_ball_advance_distance(&self) -> f32 {
-        if self.count == 0 {
-            0.0
-        } else {
-            self.total_ball_advance_distance / self.count as f32
-        }
-    }
-
-    pub fn average_lateral_centering_distance(&self) -> f32 {
-        if self.count == 0 {
-            0.0
-        } else {
-            self.total_lateral_centering_distance / self.count as f32
-        }
-    }
-}
-
 #[derive(Debug, Clone)]
 struct PendingCenterTouch {
     player: PlayerId,
@@ -117,12 +40,9 @@ struct PendingCenterTouch {
 
 #[derive(Debug, Clone, Default)]
 pub struct CenterCalculator {
-    player_stats: HashMap<PlayerId, CenterPlayerStats>,
-    team_zero_stats: CenterTeamStats,
-    team_one_stats: CenterTeamStats,
+    stats: CenterStatsAccumulator,
     events: EventStream<CenterEvent>,
     pending_touch: Option<PendingCenterTouch>,
-    current_last_center_player: Option<PlayerId>,
 }
 
 impl CenterCalculator {
@@ -131,15 +51,15 @@ impl CenterCalculator {
     }
 
     pub fn player_stats(&self) -> &HashMap<PlayerId, CenterPlayerStats> {
-        &self.player_stats
+        self.stats.player_stats()
     }
 
     pub fn team_zero_stats(&self) -> &CenterTeamStats {
-        &self.team_zero_stats
+        self.stats.team_zero_stats()
     }
 
     pub fn team_one_stats(&self) -> &CenterTeamStats {
-        &self.team_one_stats
+        self.stats.team_one_stats()
     }
 
     pub fn events(&self) -> &[CenterEvent] {
@@ -148,18 +68,6 @@ impl CenterCalculator {
 
     pub fn new_events(&self) -> &[CenterEvent] {
         self.events.new_events()
-    }
-
-    fn begin_sample(&mut self, frame: &FrameInfo) {
-        for stats in self.player_stats.values_mut() {
-            stats.is_last_center = false;
-            stats.time_since_last_center = stats
-                .last_center_time
-                .map(|time| (frame.time - time).max(0.0));
-            stats.frames_since_last_center = stats
-                .last_center_frame
-                .map(|last_frame| frame.frame_number.saturating_sub(last_frame));
-        }
     }
 
     fn center_event_for_position(
@@ -215,34 +123,7 @@ impl CenterCalculator {
     }
 
     fn record_center(&mut self, frame: &FrameInfo, event: CenterEvent) {
-        let player_stats = self.player_stats.entry(event.player.clone()).or_default();
-        player_stats.count += 1;
-        player_stats.total_ball_travel_distance += event.ball_travel_distance;
-        player_stats.total_ball_advance_distance += event.ball_advance_distance;
-        player_stats.total_lateral_centering_distance += event.lateral_centering_distance;
-        player_stats.longest_center_distance = player_stats
-            .longest_center_distance
-            .max(event.ball_travel_distance);
-        player_stats.last_center_time = Some(event.time);
-        player_stats.last_center_frame = Some(event.frame);
-        player_stats.time_since_last_center = Some((frame.time - event.time).max(0.0));
-        player_stats.frames_since_last_center =
-            Some(frame.frame_number.saturating_sub(event.frame));
-
-        let team_stats = if event.is_team_0 {
-            &mut self.team_zero_stats
-        } else {
-            &mut self.team_one_stats
-        };
-        team_stats.count += 1;
-        team_stats.total_ball_travel_distance += event.ball_travel_distance;
-        team_stats.total_ball_advance_distance += event.ball_advance_distance;
-        team_stats.total_lateral_centering_distance += event.lateral_centering_distance;
-        team_stats.longest_center_distance = team_stats
-            .longest_center_distance
-            .max(event.ball_travel_distance);
-
-        self.current_last_center_player = Some(event.player.clone());
+        self.stats.apply_event(frame, &event);
         self.events.push(event);
         self.pending_touch = None;
     }
@@ -298,10 +179,10 @@ impl CenterCalculator {
         live_play: bool,
     ) -> SubtrActorResult<()> {
         self.events.begin_update();
-        self.begin_sample(frame);
+        self.stats.begin_sample(frame);
         if !live_play {
             self.pending_touch = None;
-            self.current_last_center_player = None;
+            self.stats.clear_current_last();
             return Ok(());
         }
 
@@ -335,11 +216,7 @@ impl CenterCalculator {
             });
         }
 
-        if let Some(player_id) = self.current_last_center_player.as_ref() {
-            if let Some(stats) = self.player_stats.get_mut(player_id) {
-                stats.is_last_center = true;
-            }
-        }
+        self.stats.finish_sample();
 
         Ok(())
     }
